add recalculate money

+++ app/src/main/java/com/chico/homebuch/ui/UpdateMoneyMovingFragment.kt
@@ -8,13 +8,18 @@ import androidx.navigation.fragment.findNavController
 import com.chico.homebuch.R
 import com.chico.homebuch.constants.Const.MOVING_MONEY_KEY
 import com.chico.homebuch.database.AppDataBase
+import com.chico.homebuch.database.entity.Money
 import com.chico.homebuch.database.entity.MovingMoneyInfo
 import com.chico.homebuch.utils.launchIO
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 
 class UpdateMoneyMovingFragment : Fragment() {
 
     private val movingMoneyDao by lazy {
         AppDataBase.getInstance(requireContext())?.getMovingMoneyDao()
+    }
+    private val moneyDao by lazy {
+        AppDataBase.getInstance(requireContext())?.getMoneyDao()
     }
 
     private lateinit var incomeRb: RadioButton
@@ -68,22 +73,79 @@ class UpdateMoneyMovingFragment : Fragment() {
                 ) {
                     launchIO {
                         mMoney?.let { money ->
-                            val desc = descriptionEt.text.toString()
                             val sum = sumEt.text.toString().toDouble()
-                            val total = if (money.moneyView == 0) {
-                                money.total.plus(money.sum).minus(sum)
-                            } else {
-                                money.total.minus(money.sum).plus(sum)
-                            }
                             val view = if (incomeRb.isChecked) 1 else 0
+                            val desc = descriptionEt.text.toString()
+                            val difference = money.sum - sum
+                            val total = calculateTotalMoney(
+                                total = money.total,
+                                difference = difference,
+                                view = money.moneyView
+                            )
                             val newMMoney =
                                 MovingMoneyInfo(money.id, sum, total, view, desc, money.date)
                             movingMoneyDao?.updateMovingMoney(newMMoney)
+                            recalculateMoney(money, difference)
                         }
                     }
                     findNavController().popBackStack()
                 }
             }
+        }
+    }
+
+    private suspend fun recalculateMoney(money: MovingMoneyInfo, difference: Double) {
+        updateAfterDate(
+            date = money.date,
+            difference = difference,
+            view = money.moneyView
+        )
+        updateAccount(difference = difference, view = money.moneyView)
+    }
+
+    private fun updateAccount(difference: Double, view: Int) {
+        val totalCount = moneyDao?.getMoney()?.myMoney
+        totalCount?.let {
+            val money = calculateTotalMoney(
+                total = it,
+                difference = difference,
+                view = view
+            )
+            moneyDao?.updateMoney(Money(myMoney = money))
+        }
+    }
+
+    private suspend fun updateAfterDate(date: Long, difference: Double, view: Int) {
+        val moneyList =
+            movingMoneyDao?.getMovingMoneyAfterDate(date) ?: emptyList()
+        for (movingMoney in moneyList) {
+            val total = calculateTotalMoney(
+                total = movingMoney.total,
+                difference = difference,
+                view = view
+            )
+            val newMMoney =
+                MovingMoneyInfo(
+                    movingMoney.id,
+                    movingMoney.sum,
+                    total,
+                    movingMoney.moneyView,
+                    movingMoney.description,
+                    movingMoney.date
+                )
+            movingMoneyDao?.updateMovingMoney(newMMoney)
+        }
+    }
+
+    private fun calculateTotalMoney(
+        total: Double,
+        difference: Double,
+        view: Int
+    ): Double {
+        return if (view == 0) {
+            total + difference
+        } else {
+            total - difference
         }
     }
 
@@ -94,13 +156,30 @@ class UpdateMoneyMovingFragment : Fragment() {
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         if (item.itemId == R.id.delete_menu_item) {
-            launchIO {
-                mMoney?.let {
-                    movingMoneyDao?.deleteMovingMoney(it)
-                    findNavController().popBackStack()
+            createAlertDialog(
+                resources.getString(R.string.deletion),
+                resources.getString(R.string.are_you_sure)
+            )
+                .setPositiveButton(resources.getString(R.string.delete)) { _, _ ->
+                    launchIO {
+                        mMoney?.let {
+                            movingMoneyDao?.deleteMovingMoney(it)
+                            recalculateMoney(it, it.sum)
+                            findNavController().popBackStack()
+                        }
+                    }
                 }
-            }
+                .show()
         }
         return false
+    }
+
+    private fun createAlertDialog(title: String, message: String): MaterialAlertDialogBuilder {
+        return MaterialAlertDialogBuilder(requireContext())
+            .setTitle(title)
+            .setMessage(message)
+            .setNegativeButton(resources.getString(R.string.cancel)) { dialog, _ ->
+                dialog.cancel()
+            }
     }
 }
